Moved files around for react components
