+++ client/components/header-tabs/main.js
@@ -1,0 +1,1 @@
+main.js
