+++ client/main.scss
@@ -15,7 +15,7 @@
 @import "icons";
 
 @import "components/feed/main";
-@import "components/feed/fastft/main";
+@import "components/fastft/main";
 
 @import "components/flexipod/main";
