+++ components/fastft/main.scss
