+++ server/routes/front-page.js
@@ -1,6 +1,6 @@
 import content from '../lib/content';
-import FastFtFeed from '../../components/feed/fastft/main';
-import Feed from '../../components/feed/main';
+import {FastFtFeed} from '../../components/fastft/main';
+import {Feed} from '../../components/feed/main';
 
 module.exports = function(req, res) {
 	const useElasticSearch = res.locals.flags.elasticSearchItemGet.isSwitchedOn;
